+++ samples/vst/again/source/againsimple.cpp
@@ -355,10 +355,10 @@ tresult PLUGIN_API AGainSimple::process (ProcessData& data)
 			{
 				if (data.symbolicSampleSize == kSample32)
 					fVuPPM = processAudio<Sample32> ((Sample32**)in, (Sample32**)out, numChannels,
-						data.numSamples, gain);
+					                                 data.numSamples, gain);
 				else
-					fVuPPM = processAudio<Sample64> ((Sample64**)in, (Sample64**)out, numChannels,
-						data.numSamples, gain);
+					fVuPPM = static_cast<float> (processAudio<Sample64> (
+					    (Sample64**)in, (Sample64**)out, numChannels, data.numSamples, gain));
 			}
 		}
 	}
